+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for the Olimex ARM-USB-TINY-H JTAG device (#1586).
 - Added support for propagating `CoreStatus` to the probe in use (#1588).
 - Added PY32F0xx series targets (#1619).
+- Allow change FTDI interface through environment variable `FTDI_INTERFACE`.
 
 ## [0.18.0]